rename ctp-client.js to ctp.js and moved to parent folder.
